refactor: 각각의 DTO를 각 Entity의 api/types.ts로 이동
