+++ client/src/entities/manual-record/model/types.ts
@@ -1,0 +1,9 @@
+export interface ManualRecord {
+  id: string;
+  participantId: string;
+  value: number;
+  source: "stopwatch" | "manual" | "other";
+  status: "pending" | "approved" | "rejected";
+  note: string;
+  createdAt: Date;
+}
