+++ client/src/entities/stopwatch/api/types.ts
@@ -1,0 +1,7 @@
+export interface StopwatchDto {
+  id: string;
+  name: string;
+  startedAt: number | null;
+  stoppedAt: number | null;
+  divisionId: string | null;
+}
